refactor: updates page metadata
